+++ backend/src/routes.js
@@ -1,35 +1,8 @@
 const { Router } = require('express');
-const axios = require('axios');
-const Dev = require('./models/Dev');
+const DevController = require('./controllers/DevController');
 
 const routes = Router();
 
-routes.post('/devs', async (request, response) => 
-{
-    const { github_username, techs, latitude, longitude } = request.body;
-
-    const apiResponse = await axios.get(`https://api.github.com/users/${github_username}`);
-
-    const { name = login, avatar_url, bio } = apiResponse.data;
-
-    const techsArray = techs.split(',').map(tech => tech.trim());
-
-    const location = {
-        type: 'Point',
-        coordinates: [longitude, latitude]
-    }
-
-    const dev = await Dev.create({
-        github_username,
-        name,
-        techs: techsArray,
-        avatar_url,
-        bio,
-        location
-    });
-
-    return response.json(dev);
-});
-
+routes.post('/devs', DevController.store);
 
 module.exports = routes;
